+++ packages/react/src/Tabs/TabsButton.test.tsx
@@ -1,7 +1,6 @@
-import { fireEvent, render, screen } from '@testing-library/react'
+import { render, screen } from '@testing-library/react'
 import { createRef } from 'react'
 import { TabsButton } from './TabsButton'
-import { TabsContext } from './TabsContext'
 import '@testing-library/jest-dom'
 
 describe('Tabs button', () => {
@@ -51,26 +50,6 @@ describe('Tabs button', () => {
     const component = screen.getByRole('tab')
 
     expect(component).toHaveAttribute('aria-controls', '-panel-0')
-  })
-
-  it('calls onChange with the newly activated tab', () => {
-    const onChange = jest.fn()
-    const contextValue = {
-      activeTab: 0,
-      tabsId: 'test-tabs',
-      updateTab: onChange,
-    }
-
-    render(
-      <TabsContext.Provider value={contextValue}>
-        <TabsButton tab={1}>Tab 1</TabsButton>
-      </TabsContext.Provider>,
-    )
-
-    const button = screen.getByRole('tab', { name: 'Tab 1' })
-    fireEvent.click(button)
-
-    expect(onChange).toHaveBeenCalledWith(1)
   })
 
   it('supports ForwardRef in React', () => {
